Extract copy action into its own resource
