+++ app/controllers/concerns/template_methods.rb
@@ -1,0 +1,9 @@
+module TemplateMethods
+
+  private
+
+  def template_type(template)
+    template.customization_of.present? ? _("customisation") : _("template")
+  end
+
+end
