+++ app/helpers/org_admin/template_copies_helper.rb
@@ -1,0 +1,2 @@
+module OrgAdmin::TemplateCopiesHelper
+end
